+++ packages/components/choropleth/src/ChoroplethDiv.svelte
@@ -9,23 +9,25 @@
 	export let title;
 
 	// ChoroplethG: required
-	export let topojson;
-	export let topojsonId;
+	export let topojson = null;
+	export let topojsonId = null;
+
+	// ChoroplethG: one of the two is required
+	export let key = null;
+	export let key_alt = null;
 
 	// ChoroplethG: optional
-	export let focusedKey;
-	export let geometry;
-	export let isInteractive;
-	export let key_alt;
-	export let key;
-	export let keyToColor;
-	export let keyToColorFn;
-	export let message;
-	export let projection;
-	export let projectionFn;
-	export let projectionId;
-	export let selectedKeys;
-	export let theme;
+	export let focusedKey = null;
+	export let geometry = null;
+	export let isInteractive = false;
+	export let keyToColor = null;
+	export let keyToColorFn = null;
+	export let message = null;
+	export let projection = null;
+	export let projectionFn = null;
+	export let projectionId = null;
+	export let selectedKeys = [];
+	export let theme = null;
 
 	$: padding = padding || '10px';
 	$: headerHeight = headerHeight || '2rem';